Refactor out lnd child process code out of electron.js
